Initial Version
Version 1.3, up until now I have kept this project private on my computer, but I have decided to make this open source. Hopefully someone finds this useful. This light controller has several modes and a CLI interface with persistence in eeprom. I plan to continue development on this project, including documentation, an IO extender board (for the RJ-45 jack on the back), as well as making a board revision for more modern parts. I also want to change the coding platform from arduino IDE to PlatformIO, for ease of programming.